requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Flask-Bcrypt==1.0.1
Werkzeug==3.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
deepface==0.0.95
gunicorn==21.2.0
